Extracted factory interfaces (signatures) from main index.ts.
