+++ src/factory/signature/index.ts
@@ -1,0 +1,2 @@
+export {CombineSignature} from './combineSignature'
+export {MergeSignature} from './mergeSignature'
